feat: remove TokenField by adding balance prop to TokenFieldWithBalance
