--- src/component-library/TokenField/index.tsx
@@ -1,2 +1,0 @@
-export { TokenField } from './TokenField';
-export type { TokenFieldProps } from './TokenField';
