+++ src/component-library/TokenFieldWithBalance/TokenFieldWithBalance.stories.tsx
@@ -4,16 +4,22 @@ import { TokenFieldWithBalance, TokenFieldWithBalanceProps } from './';
 
 const Template: Story<TokenFieldWithBalanceProps> = (args) => <TokenFieldWithBalance {...args} />;
 
-const Default = Template.bind({});
-Default.args = {
+const WithBalance = Template.bind({});
+WithBalance.args = {
   tokenSymbol: 'KSM',
   valueInUSD: '100.00',
   defaultValue: '100.00', // `value`
-  balanceValue: '1000.00',
-  balanceValueInUSD: '1000.00'
+  balance: '1000.00'
 };
 
-export { Default };
+const WithoutBalance = Template.bind({});
+WithoutBalance.args = {
+  tokenSymbol: 'KSM',
+  valueInUSD: '100.00',
+  defaultValue: '100.00' // `value`
+};
+
+export { WithBalance, WithoutBalance };
 
 export default {
   title: 'Forms/TokenFieldWithBalance',
